feat: add new task keyboard shortcut
