+++ package.json
@@ -33,6 +33,7 @@
     "@tailwindcss/vite": "^4.1.18",
     "@tanstack/electric-db-collection": "^0.2.33",
     "@tanstack/react-db": "0.1.70",
+    "@tanstack/react-hotkeys": "^0.3.2",
     "@tanstack/react-query": "^5.90.5",
     "@tanstack/react-router": "^1.162.8",
     "@tanstack/react-start": "^1.162.8",
