+++ src/components/new-task-button.tsx
@@ -1,10 +1,7 @@
-import { useState, type ComponentProps } from "react";
-import { useNavigate } from "@tanstack/react-router";
-import { useLiveQuery } from "@tanstack/react-db";
+import type { ComponentProps } from "react";
 import { Loader2, Plus } from "lucide-react";
 import { Button } from "@/components/ui/button";
-import { projectsCollection, tasksCollection } from "@/lib/collections";
-import { createDesktopRunnerSession } from "@/lib/desktop-runner";
+import { useCreateNewTask } from "@/lib/use-create-new-task";
 
 type ButtonProps = ComponentProps<typeof Button>;
 
@@ -13,56 +10,14 @@ type NewTaskButtonProps = Omit<ButtonProps, "children" | "disabled" | "onClick">
 };
 
 export function NewTaskButton({ iconOnly = false, title, ...props }: NewTaskButtonProps) {
-  const navigate = useNavigate();
-  const [creating, setCreating] = useState(false);
-  const { data: projects } = useLiveQuery((query) =>
-    query.from({ p: projectsCollection }).orderBy(({ p }) => p.created_at, "asc"),
-  );
-
-  const [defaultProject] = projects;
-
-  async function handleNewTask() {
-    const repoUrl = defaultProject?.repo_url;
-    if (creating || !defaultProject || !repoUrl) {
-      return;
-    }
-
-    setCreating(true);
-
-    try {
-      const taskTitle = "New task";
-      const response = await createDesktopRunnerSession(taskTitle, repoUrl);
-      const now = Date.now();
-      const taskId = crypto.randomUUID();
-      const tx = tasksCollection.insert({
-        id: taskId,
-        organization_id: defaultProject.organization_id,
-        project_id: defaultProject.id,
-        title: taskTitle,
-        status: "open",
-        stream_id: null,
-        branch: null,
-        runner_type: response.runnerType,
-        runner_session_id: response.sessionId,
-        workspace_path: response.workspaceDirectory,
-        error: null,
-        created_at: BigInt(now),
-        updated_at: BigInt(now),
-      });
-
-      navigate({ to: "/tasks/$taskId", params: { taskId } });
-      await tx.isPersisted.promise;
-    } finally {
-      setCreating(false);
-    }
-  }
+  const { canCreateTask, createNewTask, creating } = useCreateNewTask();
 
   return (
     <Button
       type="button"
-      title={title ?? "New task"}
-      disabled={creating || !defaultProject}
-      onClick={() => void handleNewTask()}
+      title={title ?? "New task (Cmd/Ctrl+N)"}
+      disabled={!canCreateTask}
+      onClick={() => void createNewTask()}
       {...props}
     >
       {creating ? (
